+++ ext/duckdb/result.c
@@ -21,6 +21,31 @@ static VALUE to_ruby_obj_boolean(duckdb_result *result, size_t col_idx, size_t r
     return bval ? Qtrue : Qnil;
 }
 
+static VALUE to_ruby_obj_smallint(duckdb_result *result, size_t col_idx, size_t row_idx) {
+    int16_t i16val = duckdb_value_int16(result, col_idx, row_idx);
+    return INT2FIX(i16val);
+}
+
+static VALUE to_ruby_obj_integer(duckdb_result *result, size_t col_idx, size_t row_idx) {
+    int32_t i32val = duckdb_value_int32(result, col_idx, row_idx);
+    return INT2NUM(i32val);
+}
+
+static VALUE to_ruby_obj_bigint(duckdb_result *result, size_t col_idx, size_t row_idx) {
+    int64_t i64val = duckdb_value_int64(result, col_idx, row_idx);
+    return rb_int2big(i64val);
+}
+
+static VALUE to_ruby_obj_float(duckdb_result *result, size_t col_idx, size_t row_idx) {
+    float fval = duckdb_value_float(result, col_idx, row_idx);
+    return DBL2NUM(fval);
+}
+
+static VALUE to_ruby_obj_double(duckdb_result *result, size_t col_idx, size_t row_idx) {
+    double dval = duckdb_value_double(result, col_idx, row_idx);
+    return DBL2NUM(dval);
+}
+
 static VALUE to_ruby_obj(duckdb_result *result, size_t col_idx, size_t row_idx) {
     char *p;
     VALUE obj = Qnil;
@@ -30,6 +55,16 @@ static VALUE to_ruby_obj(duckdb_result *result, size_t col_idx, size_t row_idx) 
     switch(result->columns[col_idx].type) {
     case DUCKDB_TYPE_BOOLEAN:
         return to_ruby_obj_boolean(result, col_idx, row_idx);
+    case DUCKDB_TYPE_SMALLINT:
+        return to_ruby_obj_smallint(result, col_idx, row_idx);
+    case DUCKDB_TYPE_INTEGER:
+        return to_ruby_obj_integer(result, col_idx, row_idx);
+    case DUCKDB_TYPE_BIGINT:
+        return to_ruby_obj_bigint(result, col_idx, row_idx);
+    case DUCKDB_TYPE_FLOAT:
+        return to_ruby_obj_float(result, col_idx, row_idx);
+    case DUCKDB_TYPE_DOUBLE:
+        return to_ruby_obj_double(result, col_idx, row_idx);
     default:
         p = duckdb_value_varchar(result, col_idx, row_idx);
         obj = rb_str_new2(p);